Refactor the internals of Buildarr to improve maintainability
