+++ buildarr/plugins/sonarr/config/__init__.py
@@ -26,7 +26,8 @@ from typing import TYPE_CHECKING, Dict, Optional
 
 from typing_extensions import Self
 
-from buildarr.config import ConfigPlugin, NonEmptyStr, Port
+from buildarr.config import ConfigPlugin
+from buildarr.types import NonEmptyStr, Port
 
 from ..api import get_initialize_js
 from ..util import SonarrApiKey, SonarrProtocol
