+++ buildarr/plugins/sonarr/config/profiles/release.py
@@ -29,9 +29,9 @@ from typing import Any, Dict, Iterable, List, Mapping, Optional, Set, cast
 from pydantic import validator
 from typing_extensions import Self
 
-from buildarr.config import NonEmptyStr, RemoteMapEntry, TrashID
-from buildarr.config.exceptions import ConfigTrashIDNotFoundError
+from buildarr.config import ConfigTrashIDNotFoundError, RemoteMapEntry
 from buildarr.logging import plugin_logger
+from buildarr.types import NonEmptyStr, TrashID
 
 from ...api import api_delete, api_get, api_post, api_put
 from ...secrets import SonarrSecrets
