+++ buildarr/plugins/sonarr/secrets.py
@@ -24,8 +24,8 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 from urllib.parse import urlparse
 
-from buildarr.config import NonEmptyStr, Port
 from buildarr.secrets import SecretsPlugin
+from buildarr.types import NonEmptyStr, Port
 
 from .api import api_get, get_initialize_js
 from .exceptions import SonarrAPIError
